refactor(ping): store common ping payload globally
The ping operation uses the ping payload repeatedly. Hence, there is no need to constantly create and discard it in the scope of the ping function. Therefore, this change reduces GC pressure.

Additionally, the exact contents of the ping payload do not matter and, hence, can be abstracted behind a variable.